wip: start work on new skill picker
