+++ frontend/common/src/components/skills/SkillPicker/deprecated/index.ts
