+++ frontend/talentsearch/src/js/components/experienceForm/ExperienceSkills.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useIntl } from "react-intl";
 import { useFieldArray, useFormContext } from "react-hook-form";
 
-import SkillPicker from "@common/components/skills/SkillPicker";
+import SkillPicker from "@common/components/skills/SkillPicker/deprecated";
 import { notEmpty } from "@common/helpers/util";
 import type { Skill } from "../../api/generated";
