+++ frontend/talentsearch/src/js/components/skills/AddSkillsToFilter/AddSkillsToFilter.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useIntl } from "react-intl";
 
 import type { Skill } from "@common/api/generated";
-import SkillPicker from "@common/components/skills/SkillPicker";
+import SkillPicker from "@common/components/skills/SkillPicker/deprecated";
 import { useFieldArray, useFormContext } from "react-hook-form";
 import { notEmpty } from "@common/helpers/util";
